Add Test for app.Static* functions
