+++ coverage-tests/coverage.go
@@ -1,0 +1,1 @@
+package coverageTests
