+++ coverage-tests/fixtures/static/test.css
@@ -1,0 +1,1 @@
+body{font-size: 30px}
